Bug in getpermutation/BRB.D
The previous bug correction of getpermutation resulted in
integer values outside the representation of int (*1e10)
changed to 1e8 and now ok. Bug in BRB.D (when no triplet of
values was available, the diffusion cannot be computed.
Now, the function explicitly returns a NaN, instead of
relying on an unknown behaviour (diagnosed with UBSAN).